- Funkce start přesunuta z triangle.py do main.py - Opraveno volání funkce triangleOrthogonality z main.py
